Add .clang-format file with default Google style
A lot of IDE's, and editors plugins interface well with clang-tools.
This adds the file .clang-format in the top folder (the expected place).
It is populated with the default google style.

It was generated with the command:

```
clang-format -style=google -dump-config > .clang-format
```

With

```
clang-format --version
clang-format version 10.0.0
```

This should conform with the same style using `cpplint`.