EventDeduplication -> SameInputSetIsomorphicOutputs (#404)
## Changes
* Closes #341.
* Adds `"EventDeduplication"` option, which, if set to `"SameInputSetIsomorphicOutputs"`, allows one to treat multiple matches to the same set of inputs producing isomorphic outputs as one.

## Review questions
* Can you think of a better name than `"SameInputSetIsomorphicOutputs"`?

Keep in mind that another possibility for event identification would be, for example, a rule `{{0}, {0}} -> {}` starting with an input `{{0}, {0}, {0}}`. In this case, there are three events possible, but they are still the same in the sense that the states after instantiating any of these events would be identical. So, we might want to add an option for this case as well. Maybe, as an element in a list `"EventDeduplication" -> {"SameInputSetIsomorphicOutputs", ...}`. And if all possible identification are needed, `"EventDeduplication" -> All`.

## Examples
* By default, matches to every possible order of the input subsets will be turned into separate events:
```wl
In[] := WolframModel[{{a, b}, {a, c}} -> {{b, d}, {c, d}}, {{1, 2}, {1, 3}}, 
  "EventSelectionFunction" -> 
   "MultiwaySpacelike"]["ExpressionsEventsGraph", 
 VertexLabels -> Automatic]
```
<img width="399" alt="image" src="https://user-images.githubusercontent.com/1479325/93140017-bdaa8480-f6af-11ea-9339-96599bcf36b7.png">

* However, with `"EventDeduplication" -> "SameInputSetIsomorphicOutputs"` they will be identified if the outputs are the same up to the renaming of new atoms:
```wl
In[] := WolframModel[{{a, b}, {a, c}} -> {{b, d}, {c, d}}, {{1, 2}, {1, 3}}, 
  "EventSelectionFunction" -> "MultiwaySpacelike", 
  "EventDeduplication" -> "SameInputSetIsomorphicOutputs"]["ExpressionsEventsGraph", 
 VertexLabels -> Automatic]
```
<img width="259" alt="image" src="https://user-images.githubusercontent.com/1479325/93140076-d5820880-f6af-11ea-8336-e47edb8a176a.png">

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/maxitg/setreplace/404)
<!-- Reviewable:end -->