Applied comments of Max, name change to FeatureVector
Applied the comments of Max about style and code.

Fixed (hopefully every) code style issues

Changed the name of the property to FeatureVector (rightly, is not a function)

Expanded VertexDegreeQuantiles to the corner case of empty graph, returning a list of zeros.

Expanded the tests with also an empty evolution

Other minor changes (mostly in the naming of  helper functions)