Ignore Mathematica notebooks, workspace files, and specially-named (#530)
This change adds notebook files to the .gitignore list, for two reasons:

To make it much harder to deliberately or accidentally add notebooks to commits.
To make it friendlier to have a couple of "local notebooks" that are simply invisible to Git. (I use this for various purposes).
I've also added VSCode and Sublime workspace files, as well as files that contain the magic string "ignoreme.", which comes in handy when you want to 'stash' a single new file away easily without messing with git.